*)表id改为字符串类型 t_out_side_bpm_admin_personnel bpm_process_permissions bpm_process_app_application bpm_flowruninfo t_user_message t_user_message_status t_out_side_bpm_conditions_template t_bpmn_node_role_conf t_user_entrust

+++ antflow-engine/src/main/java/org/openoa/engine/bpmnconf/service/impl/BpmFlowruninfoServiceImpl.java
@@ -28,7 +28,7 @@ public class BpmFlowruninfoServiceImpl extends ServiceImpl<BpmFlowruninfoMapper,
     public void createFlowRunInfo(String entryId, String processInstance) throws Exception {
         BpmFlowruninfo flowruninfo = new BpmFlowruninfo();
         flowruninfo.setEntitykey(entryId);
-        flowruninfo.setCreateUserId( SecurityUtils.getLogInEmpIdSafe().longValue());
+        flowruninfo.setCreateUserId( SecurityUtils.getLogInEmpIdSafe());
         flowruninfo.setRuninfoid(Long.parseLong(processInstance));
         flowruninfo.setEntityclass(SecurityUtils.getLogInEmpNameSafe());
         flowruninfo.setCreateactor(SecurityUtils.getLogInEmpNameSafe());
